+++ handlr-regex/src/lib.rs
@@ -4,10 +4,3 @@ pub mod common;
 pub mod config;
 pub mod error;
 pub mod utils;
-
-pub use apps::*;
-pub use cli::*;
-pub use common::*;
-pub use config::*;
-pub use error::*;
-pub use utils::*;
