+++ handlr-regex/src/utils.rs
@@ -1,4 +1,4 @@
-use crate::Result;
+use crate::error::Result;
 pub fn notify(title: &str, msg: &str) -> Result<()> {
     std::process::Command::new("notify-send")
         .args(["-t", "10000", title, msg])
